tests for packages 1,2

+++ 02_figures-tests/src/main/java/foo/bar/Figure.java
@@ -1,6 +1,6 @@
 package foo.bar;
 
 interface Figure {
-    int area();
-    int circumference();
+    double area();
+    double circumference();
 }
