+++ 02_figures-tests/src/main/java/foo/bar/Square.java
@@ -8,10 +8,10 @@ public class Square implements Figure {
         this.a = a;
     }
 
-    public int area() {
+    public double area() {
         return a*a;
     }
-    public int circumference() {
+    public double circumference() {
         return 4*a;
     }
 }
